backend/pyproject.toml
[tool.poetry]
name = "minicampus-backend"
version = "0.1.0"
description = "CampusFlow Backend API"
authors = ["CampusFlow Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
authlib = "^1.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88